+++ .ci/logstash-run.sh
@@ -3,11 +3,17 @@ set -ex
 
 export PATH=$BUILD_DIR/gradle/bin:$PATH
 
+# CentOS 7 using curl defaults does not enable TLSv1.3
+CURL_OPTS="-k --tlsv1.2 --tls-max 1.3"
+
 wait_for_es() {
   echo "Waiting for elasticsearch to respond..."
   es_url="http://elasticsearch:9200"
+  if [[ "$SECURE_INTEGRATION" == "true" ]]; then
+    es_url="https://elasticsearch:9200"
+  fi
   count=120
-  while ! curl --silent $es_url && [[ $count -ne 0 ]]; do
+  while ! curl $CURL_OPTS -u elastic:$ELASTIC_PASSWORD --silent $es_url && [[ $count -ne 0 ]]; do
     count=$(( $count - 1 ))
     [[ $count -eq 0 ]] && return 1
     sleep 1
@@ -18,9 +24,9 @@ wait_for_es() {
 }
 
 if [[ "$INTEGRATION" != "true" ]]; then
-  bundle exec rspec -fd spec/filters -t ~integration
+  jruby -rbundler/setup -S rspec -fd -t ~integration spec/filters
 else
-  extra_tag_args="--tag integration"
+  extra_tag_args="-t integration"
   wait_for_es
-  bundle exec rspec -fd $extra_tag_args --tag es_version:$ELASTIC_STACK_VERSION spec/filters/integration
+  jruby -rbundler/setup -S rspec -fd $extra_tag_args -t es_version:$ELASTIC_STACK_VERSION spec/filters/integration
 fi